moved admin definitions to admin.py and added Term model
